requirements.txt
torch==1.12.0
torchvision==0.13.0
Pillow
accelerate==0.16.0
opencv-python
huggingface-hub==0.12.0
transformers==4.26.1
requests